+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Latest
 
 - Added all values to the Species.Category choices.
+- Added text choices for Checklist protocols.
 
 ## 0.2.1